Remove redundant django5-celery-integration-collector
The django5-celery setup now supports collector mode through the mode
toggle, so the separate integration-collector setup is no longer needed.
Its only app-level difference, a /logs endpoint that emits log messages,
is folded into django5-celery.
