+++ README.md
@@ -83,7 +83,6 @@ rake app=python/django4-celery app:up
 rake app=python/django4-wsgi app:up
 rake app=python/django5-celery app:up
 rake app=python/django5-celery-collector app:up
-rake app=python/django5-celery-integration-collector app:up
 rake app=python/fastapi app:up
 rake app=python/fastapi-databases app:up
 rake app=python/flask app:up
@@ -178,7 +177,6 @@ rake app=python/django4-celery app:bot
 rake app=python/django4-wsgi app:bot
 rake app=python/django5-celery app:bot
 rake app=python/django5-celery-collector app:bot
-rake app=python/django5-celery-integration-collector app:bot
 rake app=python/fastapi app:bot
 rake app=python/fastapi-databases app:bot
 rake app=python/flask app:bot
@@ -288,7 +286,6 @@ rake app=python/django4-celery app:restart
 rake app=python/django4-wsgi app:restart
 rake app=python/django5-celery app:restart
 rake app=python/django5-celery-collector app:restart
-rake app=python/django5-celery-integration-collector app:restart
 rake app=python/fastapi app:restart
 rake app=python/fastapi-databases app:restart
 rake app=python/flask app:restart
@@ -391,7 +388,6 @@ rake app=python/django4-celery app:bash
 rake app=python/django4-wsgi app:bash
 rake app=python/django5-celery app:bash
 rake app=python/django5-celery-collector app:bash
-rake app=python/django5-celery-integration-collector app:bash
 rake app=python/fastapi app:bash
 rake app=python/fastapi-databases app:bash
 rake app=python/flask app:bash
@@ -494,7 +490,6 @@ rake app=python/django4-celery app:console
 rake app=python/django4-wsgi app:console
 rake app=python/django5-celery app:console
 rake app=python/django5-celery-collector app:console
-rake app=python/django5-celery-integration-collector app:console
 rake app=python/fastapi app:console
 rake app=python/fastapi-databases app:console
 rake app=python/flask app:console
@@ -597,7 +592,6 @@ rake app=python/django4-celery app:diagnose
 rake app=python/django4-wsgi app:diagnose
 rake app=python/django5-celery app:diagnose
 rake app=python/django5-celery-collector app:diagnose
-rake app=python/django5-celery-integration-collector app:diagnose
 rake app=python/fastapi app:diagnose
 rake app=python/fastapi-databases app:diagnose
 rake app=python/flask app:diagnose
@@ -700,7 +694,6 @@ rake app=python/django4-celery app:tail:appsignal
 rake app=python/django4-wsgi app:tail:appsignal
 rake app=python/django5-celery app:tail:appsignal
 rake app=python/django5-celery-collector app:tail:appsignal
-rake app=python/django5-celery-integration-collector app:tail:appsignal
 rake app=python/fastapi app:tail:appsignal
 rake app=python/fastapi-databases app:tail:appsignal
 rake app=python/flask app:tail:appsignal
