--- python/django5-celery-integration-collector/app/.gitignore
@@ -1,2 +1,0 @@
-__pycache__
-tmp/
