--- python/django5-celery-integration-collector/app/__appsignal__.py
@@ -1,6 +1,0 @@
-from appsignal import Appsignal
-
-appsignal = Appsignal(
-    active=True,
-    environment="development"
-)
