--- python/django5-celery-integration-collector/app/appsignal_python_opentelemetry/templates/home.html
@@ -1,18 +1,0 @@
-<h1>Python OpenTelemetry app</h1>
-
-<ul>
-  <li><a href="/custom">/custom: Very customized route</a></li>
-  <li><a href="/slow">/slow: Trigger a slow request</a></li>
-  <li><a href="/slow_queue">/slow_queue: Trigger a slow background task</a></li>
-  <li><a href="/slow_queue_inline">/slow_queue_inline: Trigger a slow background task inline</a></li>
-  <li><a href="/error">/error: Trigger an error</a></li>
-  <li><a href="/custom_error">/custom_error: Trigger a custom error</a></li>
-  <li><a href="/error_queue">/error_queue: Trigger an error background task</a></li>
-  <li><a href="/error_queue_inline">/error_queue_inline: Trigger an error background task inline</a></li>
-  <li><a href="/metrics">/metrics: Send metrics</a></li>
-  <li><a href="/make_request">/make_request: Make a request to www.appsignal.com</a></li>
-  <li><a href="/custom_instrumentation">/custom_instrumentation: Report custom instrumentation</a></li>
-  <li><a href="/logs">/logs: Report logs</a></li>
-  <li><a href="/blog">/blog: Query and list blog posts</a></li>
-  <li><a href="/blog/create">/blog/create: Create and show a new blog post</a></li>
-</ul>
