--- python/django5-celery-integration-collector/app/appsignal_python_opentelemetry/wsgi.py
@@ -1,16 +1,0 @@
-"""
-WSGI config for appsignal_python_opentelemetry project.
-
-It exposes the WSGI callable as a module-level variable named ``application``.
-
-For more information on this file, see
-https://docs.djangoproject.com/en/4.1/howto/deployment/wsgi/
-"""
-
-import os
-
-from django.core.wsgi import get_wsgi_application
-
-os.environ.setdefault('DJANGO_SETTINGS_MODULE', 'appsignal_python_opentelemetry.settings')
-
-application = get_wsgi_application()
