--- python/django5-celery-integration-collector/app/blog/apps.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-class BlogConfig(AppConfig):
-    default_auto_field = 'django.db.models.BigAutoField'
-    name = 'blog'
