--- python/django5-celery-integration-collector/app/blog/models.py
@@ -1,5 +1,0 @@
-from django.db import models
-
-class Post(models.Model):
-    title = models.CharField(max_length=30)
-    body = models.CharField(max_length=10000)
