--- python/django5-celery-integration-collector/app/blog/templates/show.html
@@ -1,3 +1,0 @@
-<h1>{{ post.title }}</h1>
-
-<p>{{ post.body }}</p>
