--- python/django5-celery-integration-collector/app/processmon.toml
@@ -1,21 +1,0 @@
-[[paths_to_watch]]
-path = "/app"
-
-[processes.django]
-command = "python3"
-args = [
-  "manage.py",
-  "runserver",
-  "0.0.0.0:4001"
-]
-working_dir = "/app"
-
-[processes.celery]
-command = "celery"
-args = [
-  "--app=tasks",
-  "worker",
-  "--loglevel=INFO",
-  "--queues=high-priority,low-priority"
-]
-working_dir = "/app"
