--- python/django5-celery-integration-collector/app/requirements.txt
@@ -1,16 +1,0 @@
-celery
-appsignal
-django==5.1.8
-Jinja2
-psycopg2
-redis
-requests
-opentelemetry-api
-opentelemetry-sdk
-opentelemetry-exporter-otlp-proto-http
-opentelemetry-instrumentation-celery
-opentelemetry-instrumentation-django==0.53b0
-opentelemetry-instrumentation-psycopg2
-opentelemetry-instrumentation-redis
-opentelemetry-instrumentation-jinja2
-opentelemetry-instrumentation-requests
