--- python/django5-celery-integration-collector/commands/run
@@ -1,23 +1,0 @@
-#!/bin/bash
-
-set -eu
-
-pip3 install --upgrade pip
-
-cd /integration
-
-pip3 install hatch "click<8.3.0"
-hatch clean
-hatch run build:me
-
-cd /app
-
-echo "Installing dependencies"
-pip3 install -r requirements.txt
-pip3 install /integration/dist/* --force-reinstall
-
-echo "Running migrations"
-python3 manage.py migrate
-
-echo "Starting app"
-/commands/processmon processmon.toml
