--- python/django5-celery-integration-collector/postgres.env
@@ -1,4 +1,0 @@
-POSTGRES_USER=app
-POSTGRES_PASSWORD=password
-POSTGRES_DB=app_db
-POSTGRES_HOST=postgres
