+++ python/django5-celery/app/appsignal_python_opentelemetry/templates/home.html
@@ -12,6 +12,7 @@
   <li><a href="/metrics">/metrics: Send metrics</a></li>
   <li><a href="/make_request">/make_request: Make a request to www.appsignal.com</a></li>
   <li><a href="/custom_instrumentation">/custom_instrumentation: Report custom instrumentation</a></li>
+  <li><a href="/logs">/logs: Report logs</a></li>
   <li><a href="/blog">/blog: Query and list blog posts</a></li>
   <li><a href="/blog/create">/blog/create: Create and show a new blog post</a></li>
 </ul>
